change types file to .d
